Implemented read and write for specific bits

+++ src/test.rs
@@ -11,17 +11,17 @@ pub mod test {
         let b2 = Bitfield::<u8>::from_num(11); // 1011
         assert!(b2.contains(&b1));
 
-        let t1 = Bitfield::<u8>::from_num(156); // 1001 1100
-        let t2 = Bitfield::<u8>::from_num(20); // 0001 0100
+        let t1 = Bitfield::<u32>::from_num(156); // 1001 1100
+        let t2 = Bitfield::<u32>::from_num(20); // 0001 0100
 
-        let c1 = Bitfield::<u8>::from_num(0); // 0000 0000
-        let c2 = Bitfield::<u8>::from_num(0); // 0000 0000
+        let c1 = Bitfield::<usize>::from_num(0); // 0000 0000
+        let c2 = Bitfield::<usize>::from_num(0); // 0000 0000
 
         let y1 = Bitfield::<u8>::from_num(0); // 0000 0000
         let y2 = Bitfield::<u8>::from_num(1); // 0000 0001
         let y3 = Bitfield::<u8>::from_num(3); // 0000 0011
         assert!(t1.contains(&t2));
-        assert!(t1.remove(&Bitfield::<u8>::from_num(156)).unwrap().empty());
+        assert!(t1.remove(&Bitfield::<u32>::from_num(156)).unwrap().empty());
         assert!(!c1.contains(&c2));
         assert!(!y2.contains(&y1));
         assert!(!y1.contains(&y2));
